net/bnxt: fix to reset Rx next consumer index
In bnxt_init_one_rx_ring(), reset this variable internal to the driver
ring to 0, so that there is no scope for any false mismatch errors being
flagged on traffic resumption. Typically, this would show up in a port
stop/start sequence as then the HW would start afresh, but this driver
internal variable would still point to a stale value.

Fixes: 03c8f2fe111c ("net/bnxt: detect bad opaque in Rx completion")
Cc: stable@dpdk.org

Signed-off-by: Somnath Kotur <somnath.kotur@broadcom.com>
Reviewed-by: Kalesh AP <kalesh-anakkur.purayil@broadcom.com>
Reviewed-by: Ajit Khaparde <ajit.khaparde@broadcom.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>