eventdev: rx-adapter: improve enqueue buffer to circular buffer
v2:
* Fixed typo in commit message
* changed subject line

v1:
RX adapter uses memove() to move unprocessed events to the beginning of
the packet enqueue buffer. The use memmove() was found to consume good
amount of CPU cycles (about 20%).

This patch removes the use of memove() while implementing a circular
buffer to avoid copying of data. With this change RX adapter is able
to fill the buffer of 16384 events.

Signed-off-by: Ganapati Kundapura <ganapati.kundapura@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>