net/ixgbe: fix vf mac remains
In the following two scenarios, the mac address residual problem
will occur:
1) In the VF hard pass-through environment, after adding the mac
address, the process restarts, and the previously added mac is
still valid;
2) In the vf hard pass-through environment, after the mac address
is issued, the port will start/stop, and the previously added mac
is still valid;

The patch clears the mac address at the start and stop of the vf
port. After the start is cleared, the rte framework will restore
the mac addition to solve the problem of residual mac addresses.

Fixes: af75078fece3 ("first public release")
Cc: stable@dpdk.org

Signed-off-by: Qiming Chen <chenqiming_huawei@163.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>