net/ice: refine flow priority support in PF
The usage of priority is converse in pipeline mode and
non-pipeline mode. Refine attribute priority support of flow
filter in PF driver. When priority is 0, rules are created in
switch filter first and FDIR is used as backup. When priority
is 1, rules are all created in switch filter. Other filters
don't support priority 1. Value 0 denotes higher priority.

Signed-off-by: Yuying Zhang <yuying.zhang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>