lib: pdump is not supported on Windows
The original version of the pdump library was building on
Windows, but it was useless since the pdump utility was not being
built.

The new version of pdump with filtering now has dependency
on bpf. But bpf library is not available on Windows.

For now, just stop trying to build pdump on Windows.
Eventually, bpf library, pdump library, dumpcap tool,
and pdump tool can be converted to work on Windows.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>

Cc: Dmitry Kozlyuk <dmitry.kozliuk@gmail.com>
Cc: Narcisa Ana Maria Vasile <navasile@linux.microsoft.com>
Cc: Dmitry Malloy <dmitrym@microsoft.com>
Cc: Pallavi Kadam <pallavi.kadam@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>