net/iavf: fix frequent command interaction leads to high cpu
There is currently a scenario test, which will continuously obtain port
statistics, causing the CPU usage to soar, which does not meet the
demand. After positioning analysis, it is found that the vf and pf
command interaction is completed through the iavf_execute_vf_cmd function.
After the message is sent, it needs to wait for the interrupt thread to
obtain the response from the PF. For the data, the rte_delay_ms interface
is used here to wait, but the CPU will not be released during the waiting
period of this interface, which will cause the statistics to keep
occupying the CPU. This is also the root cause of the soaring cpu.
The command interaction should belong to the control plane, and there will
not be too high requirements for performance. It is recommended to wait
for the interface iavf_msec_delay to complete without taking up the CPU
time.

Fixes: 22b123a36d07 ("net/avf: initialize PMD")
Cc: stable@dpdk.org

Signed-off-by: Qiming Chen <chenqiming_huawei@163.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>