app/testpmd: fix statistics in multiple process
Currently, 'clear port stats all' in secondary will clear stats in PMD,
and also update stats which APP stores in 'ports[]' array in secondary
process, note that, that in primary process remains unchanged. So, when
'show fwd stats all' in primary process, stats in PMD may be less than
stats which APP stores in 'ports[]' array(the stats is different with
that in secondary). So forward statistics(stats in PMD minus stats
which APP stores in 'ports[]' array) will be wrong. Like this:
---------------------- Forward statistics for port 0  --------------
RX-packets: 18446744073703120168 RX-dropped: 18446744073704076766
RX-total: 18446744073697645318
TX-packets: 18446744073703122216 TX-dropped: 0
TX-total: 18446744073703122216
--------------------------------------------------------------------
Stats in PMD are shared between multiple processes, but stats which APP
stores have their own copies in multiple processes. And this will
result in bugs.

This patch will fix it by creating shared memory to store last stats for
multiple and secondary process in testpmd.

Fixes: 184de26c78d0 ("app/testpmd: support multi-process")

Signed-off-by: Min Hu (Connor) <humin29@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>